refactor: перенести всю логику в UserService
