+++ src/main/java/ru/practicum/shareit/user/User.java
@@ -9,9 +9,4 @@ public class User {
     private long id;
     private String name;
     private String email;
-
-    public User(String name, String email) {
-        this.name = name;
-        this.email = email;
-    }
 }
