+++ src/main/java/ru/practicum/shareit/user/UserMapper.java
@@ -9,6 +9,6 @@ public class UserMapper {
     }
 
     public static User toUser(UserDto userDto) {
-        return new User(userDto.getName(), userDto.getEmail());
+        return new User(userDto.getId(), userDto.getName(), userDto.getEmail());
     }
 }
